feat: matrix type picker
