+++ lib/widgets/device_grid/desktop/desktop_sidebar.dart
@@ -403,131 +403,3 @@ class _DesktopDeviceSelectorTileState extends State<DesktopDeviceSelectorTile> {
     );
   }
 }
-
-class AddExternalStreamDialog extends StatefulWidget {
-  const AddExternalStreamDialog({super.key});
-
-  static Future<void> show(BuildContext context) {
-    return showDialog(
-      context: context,
-      builder: (context) => const AddExternalStreamDialog(),
-    );
-  }
-
-  static void addStream(BuildContext context, String url, [String? name]) {
-    final loc = AppLocalizations.of(context);
-    AppLocalizations.localizationsDelegates;
-    final device = Device.dump(
-      name: name ?? loc.externalStream,
-      url: url,
-      id: const Uuid().v4().hashCode,
-    )..server = Server.dump(name: url);
-
-    final view = context.read<DesktopViewProvider>();
-    final layout = view.layouts
-        .firstWhereOrNull((layout) => layout.name == loc.externalStream);
-    if (layout == null) {
-      view.addLayout(Layout(name: loc.externalStream, devices: [device]));
-    } else {
-      view.add(device, layout);
-    }
-
-    view.updateCurrentLayout(
-      view.layouts.indexOf(
-        view.layouts.firstWhere((layout) => layout.name == loc.externalStream),
-      ),
-    );
-  }
-
-  @override
-  State<AddExternalStreamDialog> createState() =>
-      _AddExternalStreamDialogState();
-}
-
-class _AddExternalStreamDialogState extends State<AddExternalStreamDialog> {
-  final _formKey = GlobalKey<FormState>();
-  final nameController = TextEditingController();
-  final urlController = TextEditingController();
-
-  @override
-  void dispose() {
-    nameController.dispose();
-    urlController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final loc = AppLocalizations.of(context);
-    return AlertDialog(
-      title: Text(loc.addExternalStream),
-      content: ConstrainedBox(
-        constraints: BoxConstraints(
-          minWidth: MediaQuery.sizeOf(context).width * 0.425,
-        ),
-        child: Form(
-          key: _formKey,
-          child: Row(
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Expanded(
-                child: TextFormField(
-                  autofocus: true,
-                  controller: nameController,
-                  decoration: InputDecoration(label: Text(loc.streamName)),
-                  textInputAction: TextInputAction.next,
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return loc.streamNameRequired;
-                    }
-                    return null;
-                  },
-                ),
-              ),
-              const SizedBox(width: 16.0),
-              Expanded(
-                flex: 2,
-                child: TextFormField(
-                  controller: urlController,
-                  decoration: InputDecoration(label: Text(loc.streamURL)),
-                  textInputAction: TextInputAction.done,
-                  onFieldSubmitted: (_) => _finish(),
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return loc.streamNameRequired;
-                    } else if (Uri.tryParse(value) == null) {
-                      return loc.streamURLNotValid;
-                    }
-
-                    return null;
-                  },
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-      actions: [
-        FilledButton(
-          onPressed: _finish,
-          child: Text(loc.finish),
-        ),
-      ],
-    );
-  }
-
-  Future<void> _finish() async {
-    if (_formKey.currentState == null || !_formKey.currentState!.validate()) {
-      return;
-    }
-
-    AddExternalStreamDialog.addStream(
-      context,
-      urlController.text,
-      nameController.text,
-    );
-
-    Navigator.of(context).pop();
-  }
-}
