+++ lib/widgets/device_grid/device_grid.dart
@@ -24,7 +24,6 @@ import 'package:animations/animations.dart';
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:bluecherry_client/models/device.dart';
 import 'package:bluecherry_client/models/layout.dart';
-import 'package:bluecherry_client/models/server.dart';
 import 'package:bluecherry_client/providers/desktop_view_provider.dart';
 import 'package:bluecherry_client/providers/mobile_view_provider.dart';
 import 'package:bluecherry_client/providers/server_provider.dart';
@@ -36,6 +35,7 @@ import 'package:bluecherry_client/utils/theme.dart';
 import 'package:bluecherry_client/utils/video_player.dart';
 import 'package:bluecherry_client/utils/window.dart';
 import 'package:bluecherry_client/widgets/collapsable_sidebar.dart';
+import 'package:bluecherry_client/widgets/device_grid/desktop/external_stream.dart';
 import 'package:bluecherry_client/widgets/device_grid/desktop/layout_manager.dart';
 import 'package:bluecherry_client/widgets/device_grid/desktop/multicast_view.dart';
 import 'package:bluecherry_client/widgets/device_grid/mobile/device_view.dart';
@@ -53,7 +53,6 @@ import 'package:flutter/services.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:provider/provider.dart';
 import 'package:unity_video_player/unity_video_player.dart';
-import 'package:uuid/uuid.dart';
 
 part 'desktop/desktop_device_grid.dart';
 part 'desktop/desktop_sidebar.dart';
